Moves with effects specific to mons: Lick
Managed to implement Lick, yay, which checks if the user is Lickitung and changes the secondary effect chance to 100%. We also now have a battlescript function that can check a mon's species and jump accordingly. Useful, I hope!

Also, fixed a bug due to my rearrangement of status IDS :P