ESSG 2 - printing multiple stats to one string
Managed to make it work - used a cheat and some clever usage of gBattleTextBuff1 to store all data needed for string printing at once and load it in a function in battle_messages.c. Works well on Tickle and some test scenarios. To do: every other move that boosts stats! Every other call to ChangeStatBuffs and calls to abilities that prevent stat drops