Refactored all "absorbing" abilties (sans Earth Eater)
To do: a few specific functions in the AI that are a little fiddlier. Left aside Earth Eater for now since ground immunity is also an attribute of Levitate - so would like to be able to handle both. Wind Rider is possible if we define a "Flags" substruct in MovesInfo that we can also use as a substruct in AbilityEffect. This would expand the effect to also include Bulletproof and Soundproof, as well as enabling more flag-specific functionality such as Tough Claws and Long Reach.